requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
click>=8.0.0
python-dotenv>=0.19.0
requests>=2.31.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0 